Add a QuerySet type to make representing Django-style set queries easier
The QuerySet type that handles the desired semantics of `state()` and
`state_not()` (similarly `health()` and `health_not()`). The standard
REST pattern which this mimics is not entirely regular: if you do not
specify any values, all values are included; once you specify a value,
only that value is included, but specifying further values expands the
accepted set. The routine `include()` is intended to back calls like
`state()` on a JobsBuilder, and `exclude()` to back calls like
`state_not()`.

To be eligible for inclusion in a QuerySet, the item type must
implement the QuerySetMember trait, which allows the queryset to find
all possible values for the type. This only really makes sense for
simple enums. The need for all values stems from the need to elide
the field entirely from the query if no filtering has ended up being
requested, and because `exclude()` needs to know positively which
values are included.

An default initialized QuerySet provides no filtering, so need not be
wrapped in an Option.