Rename LavaMock to Server
This was previously the central object in the API, and it's still
usable as such (in a more flexible mode). But we're about to add a
simplified and more streamlined central object, so we need to rename
this one. It functions as a wiremock server object, so Server makes
sense as the new name.

Signed-off-by: Ed Smith <ed.smith@collabora.com>