Add a job results API for querying LAVA jobs
This provides a byte stream of the junit test results for a completed
LAVA job, using the standard REST junit endpoint that LAVA provides
in v0.2.

The reason to provide a raw bytestream, rather than parsed Junit data,
is the difficulty of choosing any particular junit crate for rust at
the present: none of them are comprehensive, and it really depends
what you want to do with the data which you should pick. It seems
safer for now to avoid having any policy here.

Signed-off-by: Ed Smith <ed.smith@collabora.com>