Permit deserializing Metadata
It's not clear needing to do this represents a design failure or not.
We opted to store the metadata for the TestCases instances in the mock
as strings (even though we have a Metadata type, that we in fact use
to generate that string).

We've now run into the issue that we actually need to know some of the
information in that metadata in order to populate the junit reports that
we would like to return - the failure type and message only exist in the
TestCase Metadata.

There's a more extensive rework possible here where we switch to storing
the metadata as structured, but I believe that this wasn't done because
of the interaction with django - to maintain compatibility with LAVA we
have to treat it as a string, which requires weird carve-outs (i.e. some
kind of serialize-on-demand behaviour when processing queries).

Rather than doing that, the next simplest thing is just to read back
the necessary data from the metadata by deserializing it when we need
it for Junit.  This commit introduces the necessary parts to perform
that deserialization.

Signed-off-by: Ed Smith <ed.smith@collabora.com>