Add a test for junit support to the lava-api
This is a simple conversion of one of the TestCase unit tests, except
that now we are using the data stored in the mock to double check the
contents of a parsed junit report.

This once again uses junit-parser, but only as a dev-dependency, to
simplify the determination of the contents of the generated report.

Signed-off-by: Ed Smith <ed.smith@collabora.com>