Add a basic test for mock junit support
This just retrieves four test cases and validates that they have the
expected names and durations. The output is parsed with junit-parser
(a dev-dependency) to limit sensitivity to the surface form of the XML.

This is just a quick smoke test, and not comprehensive; for one thing
all the tests have a success status, and for another there's no check
that the form matches that provided by LAVA for a real report. These
are both matters of time and convenience, in part because serde_xml
has a much more limited set of features than serde_json or serde_yaml.

Signed-off-by: Ed Smith <ed.smith@collabora.com>