Add support for junit to lava-api-mock
This adds the junit results endpoint for each job to the mocked
API. Rather than introduce new data to back this endpoint, it's driven
by the existing data for TestCases and TestSuites which already exists
in the shared state.

The major wrinkle here is the need to parse the TestCase metadata in
order to report the message and failure type in the generated junit.

Signed-off-by: Ed Smith <ed.smith@collabora.com>